Initial addition of using the information in the perfSONAR Topology Service when doing pathfinding.
It adds some minor changes to common code areas:

1. adds a lookupElement to the bss topology, domain, node, port and link elements.
  - This function will search for a given identifier in the Object requested
    and if it doesn't match, will check in the Objects children.

2. Adds fully-qualified identifiers for the id parsed along with all parent
identifiers to the results of the URNParser. A given set of parsing results
might have domainFQID, nodeFQID, portFQID and linkFQID. If a link identifier
were split up, it would include the FQIDs for all its parent elements. 

3. Now, multiple pathfinders can be specified in the file by separating them with commas. e.g.

perfsonar,database,terce

It will then go through each of the pathfinders trying to find a path through
the network. The goal is to have the static route selection and the perfsonar
path section pathfinders live side by side so users can static setup a route,
but also use the more dynamic perfsonar approach.



git-svn-id: https://oscars.es.net/repos/oscars/trunk@6963 c75a6b91-02ef-0310-b74a-e20389c9c41d