Added prelim capabilities for alternate intradomain path calculation
The alternate path is to be used for resiliency. We compute it by taking the nodes & ports used in the primary path and assigning a much higher cost to them, and then computing the shortest path in the modified topology graph.

Currently we don't do anything except log the resulting ERO to the debug log with this information. 

I used a bunch of "magic" numbers to multiply the trafficEngineeringMetric value for node->port and link->remoteLink edges... these should probably become either properties or user parameters.

git-svn-id: https://oscars.es.net/repos/oscars/trunk@7133 c75a6b91-02ef-0310-b74a-e20389c9c41d