Update PathElem.java to keep PathElemParams in a HashMap. The key to the HashMap is the concatenation of swcap and type. Right now since only two params are defined, vlanRangeAvailability and suggestedVlan, its probably overkill to have swcap as part of the key but it might be useful in the future to guarantee uniqueness.
Also added classes with constants for th PathElemParam swcap and type fields.

Added a copyPathElemParams method that copies just the parameters of a PathElem. made it separate because it will be useful in VlanMapFilter and finalizeVlanTags for copying vlanRanges between the local and interdomain path as their updated. 

Finally, used above method to copy PathElemParams in the copyPathElem() method. 


git-svn-id: https://oscars.es.net/repos/oscars/trunk@7399 c75a6b91-02ef-0310-b74a-e20389c9c41d