Notify call now works in NotificationBroker. All calls are now converted. The only work that remains is the Site checks used to determine when some users are allowed to see a Notification. I also am going to do some clean-up and minor reorganization under the notifybroker directory.
You will notice that the NotificationBroker Core is completely Axis2 free. In 0.4 we used an Axis2 OMElement to represent an opaque message to be forwarded (opaque because the NB doesn't care if its an IDC event or something else). Unfortunately it turns out the Axiom OMElement is not Serializable. This means it can't be passed via RMI to the core. I had to convert the OMElement to a JDOM type and instead of converting it back I made a JDOM HttpClient. Since we have never signed Notify messages this works quite nicely. I did some limited testing with the new NB and the 0.4 OSCARS and the messages appear to work fine.


git-svn-id: https://oscars.es.net/repos/oscars/trunk@7582 c75a6b91-02ef-0310-b74a-e20389c9c41d