Fixes for some issues when working with Junipers:
1. The GRI was used to name elements when creating them. JunOS limits certain
identifiers to 32 characters in length. One part of this commit changes how the
JunOS entity names are created to limit it to 32-characters.

If the GRI follows the normal "[domain]-[number]" pattern, the domain portion
will be truncated to ensure that the resulting id is under 32 characters.  If
it's some other form, the beginning of the GRI is removed when creating the new
id to ensure that it is 32 characters or less.

2. When the statusLSP function was called, the router id was used directly as
the address for the router. This changes the code to lookup the node address,
and use that instead of the router's id.



git-svn-id: https://oscars.es.net/repos/oscars/trunk@8308 c75a6b91-02ef-0310-b74a-e20389c9c41d