We were seeing an issue where the site administrator was unable to cancel or modify their site circuits. The error was something to the effect of "userConstraint must be set if institutionConstraint is not set".
The getConstrainedResv function had the above error condition so that the query
wouldn't hit a race condition where the "link" object for a given hop didn't
yet exist. I'm not positive why the race condition exists, but when we removed
the check, we started hitting that race condition.

This hack adds an "isCancelOrModify" boolean to the getConstrainedResv method,
and updates each caller to specify whether or not they are calling as part of a
cancel or modify function. Not pretty by any stretch, but it seems to work.



git-svn-id: https://oscars.es.net/repos/oscars/trunk@8416 c75a6b91-02ef-0310-b74a-e20389c9c41d