Consistify the version of log4j across modules as much as possible. Previously the various modules were using three different versions.
We define a new ${log4j.version} property in the top-level pom.xml
and use that definition in the individule modules' pom.xml files.
We standardize on log4j 1.2.17, which is as of this writing the
most recent log4j on the 1.2 release train.

nsibridge doesn't inherit the properties from the top-level, so we
leave that one alone (it was using 1.2.17 already).



git-svn-id: https://oscars.es.net/repos/oscars/trunk@10740 c75a6b91-02ef-0310-b74a-e20389c9c41d